server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asgiref>=3.11.0",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.6.0",
    "fastapi-mail>=1.6.0",
    "fastapi[standard]>=0.122.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.27",
]

[tool.fastapi]
entrypoint = "src.main:app"

[dependency-groups]
dev = ["ruff>=0.14.8"]

[tool.ruff]
line-length = 88
exclude = [
    ".venv/",
    "venv/",
    "env/",
    "__pypackages__/",
    "__pycache__/",
    "migrations/",
]
fix = true
indent-width = 4


[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
